fix モデル名変更
