+++ app/fast/static/text/music_list_nega.txt
@@ -1,0 +1,7 @@
+シ ド レ シ ラ ファ ラ シ ラ ファ ラ ド ラ ファ レ ド レ
+シ ド ファ ラ ソ ファ ミ レ ド ファ シ ド ファ ミ レ ド レ
+ミ ラ シ レ ド シ ラ ド レ ラ ミ ド ファ ソ ミ ファ シ
+ソ シ ミ ソ ファ ミ レ ミ レ ド シ レ ド ラ ソ ラ
+ファ ミ レ ファ ラ シ レ ド シ レ ファ ラ ド シ ラ ソ ファ ラ ソ
+レ ソ レ ソ ド ソ ラ ソ ソ ファ ソ レ ファ レ ソ シ ド レ レ ソ
+ミ ラ シ ド シ ラ ミ ミ ラ シ #ド ソ ラ ファ レ #ド シ ド シ ラ ミ
