+++ app/fast/static/text/music_list_posi.txt
@@ -1,0 +1,10 @@
+ド #ド シ ソ ミ ラ ファ レ ソ ミ ド レ ミ ファ シ ド レ ミ ファ ソ
+シ ソ ミ ファ ソ ド ラ ファ #ド シ ラ ソ #ド ソ ファ ラ ド シ ソ レ ソ
+ド #ド ミ レ ソ ソ ファ ファ ラ ラ ド #ド シ シ ソ ソ レ レ ミ ミ
+ミ ソ ミ ド ファ ラ ファ ド ミ ソ ミ #ド シ ソ ファ レ
+ミ ソ #ド シ ラ ファ レ ソ ド ミ ファ レ ソ シ レ ソ
+ド ミ ソ #ド ソ ミ ファ ラ #ド ファ ド ラ ソ シ レ ソ レ シ ド ソ ド ミ ド ソ
+ソ シ ラ レ ソ ラ シ ソ シ ソ ラ レ レ ラ シ ソ
+ミ #ド シ ラ ソ ミ レ ド シ レ ソ シ ド ミ ド ソ ド ミ ファ ソ ラ レ ミ ファ ソ シ レ ド ミ ソ ド
+ソ ミ ド ラ ファ ミ レ ソ ファ ミ ソ ミ ド ラ ファ ミ レ ソ ミ #ド
+ソ シ ソ レ シ レ ソ シ ラ #ド ラ ファ レ ファ ラ ド シ レ シ ソ レ ソ レ ソ シ レ ド レ ド ラ ファ ラ ファ レ
